+++ app/src/main/java/com/paulohenriquesg/fahrenheit/api/MediaProgressRequest.kt
@@ -1,9 +1,9 @@
 package com.paulohenriquesg.fahrenheit.api
 
 data class MediaProgressRequest(
-    val duration: Float? = null,
-    val progress: Float? = null,
-    val currentTime: Float? = null,
+    val duration: Double? = null,
+    val progress: Double? = null,
+    val currentTime: Double? = null,
     val isFinished: Boolean = false,
     val hideFromContinueListening: Boolean = false,
     val finishedAt: Long? = null,
